Connection  Chooser changes to support different databases

+++ SnailbirdData/Models/Post.cs
@@ -8,7 +8,7 @@ namespace SnailbirdData.Models
 {
     public abstract class Post : DataAccess.IModel
     {
-        public int ID { get; set; } = default!;
+        public long ID { get; set; } = default!;
         public string Title { get; set; } = default!;
         public DateTime PostDate { get; set; } = default!;
     }
